refactor: move implementations to inner classes, and use marray for enumtable
